+++ doc/sphinx/changelog.rst
@@ -9,6 +9,8 @@ performed stress tests and code audit DAS servers.
 
 - 0.7.X
 
+  - calculate run duration from RR output
+  - add conddb map into cms_maps
   - add initial support for search without DAS keywords
   - apply unique filter permanently for output results
   - add help cards to front web page to help users get use with DAS syntax